MNT move an orphaned vendored file around
File scipy/misc/pilutil.py has been removed from Scipy 1.3.0,
see scipy/scipy#9325.

Move _pilutil.py from sklearn/externals, which is the location
for vendored files, to sklearn/utils, which is a location for
locally maintained files, including orphan vendored files like
this one.
